backend/requirements.txt
# FastAPI backend requirements
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
pydantic==2.5.3

# OpenAI/LLM integration
openai>=1.50.0

# Data warehouse connections
duckdb==1.0.0
boto3==1.34.34
pandas==2.1.4
pyarrow==15.0.0

# CORS for frontend
python-multipart==0.0.6

